requirements.txt
requests
python-dotenv
numpy
pandas
schedule
